requirements.txt
flask==0.9
flask-login
flask-openid
sqlalchemy==0.7.9
flask-sqlalchemy==0.16
sqlalchemy-migrate
flask-whooshalchemy
flask-wtf==0.8.4
pytz==2013b
flask-babel==0.8
flup
lamson 
chardet 
flask-mail